+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 #### Added
 - Ignore and skip tables in automatic populator command
+- Add test combination foreign key and self foreign key to test class table analyzer
 
 ### [1.1.0] - 2021-05-26
 #### Added